+++ numba_mlir/numba_mlir/mlir/gpu_runtime.py
@@ -87,7 +87,3 @@ if IS_GPU_RUNTIME_AVAILABLE:
 
     _register_funcs()
     del _register_funcs
-
-    get_device_caps_addr = int(
-        ctypes.cast(runtime_lib.gpuxGetDeviceCapabilities, ctypes.c_void_p).value
-    )
