+++ numba_mlir_gpu_common/GpuCommon.hpp
@@ -27,13 +27,6 @@ public:
   virtual sycl::queue *getQueue() = 0;
 };
 
-struct OffloadDeviceCapabilities {
-  uint16_t spirvMajorVersion;
-  uint16_t spirvMinorVersion;
-  bool hasFP16;
-  bool hasFP64;
-};
-
 enum class GpuAllocType { Device = 0, Shared = 1, Local = 2 };
 
 enum class GpuParamType : int32_t {
